update Accept/Reject Payment Endpoint
