+++ routes/admins.js
@@ -2,5 +2,6 @@ const express = require("express");
 const Router = express.Router();
 const AdminController = require("../controllers/admins");
 
+Router.patch("/status",AdminController.updateStatusByAppointmentID);
 Router.post("/create-admin", AdminController.createAdmin);
 module.exports = Router;
